Define ``BulletStyle`` for controlling bullet style
Using ``False`` (but not ``True``), ``None``, and ``MSO_NUMBERED_BULLET_STYLE`` felt a bit messy.

``True`` (as in, use a default bullet) would be nice, but there is currently no way to infer what bullet to use. From my understanding, determining the bullet would requireus to parse the slide master and layout. Setting a "global" default could cause unexpected behaviour, and it is better that users explicitly choose a bullet.